+++ lib/asn1sequence.dart
@@ -16,7 +16,7 @@ class ASN1Sequence extends ASN1Object {
   /// we ignore any remaining bytes.
   ///
   ASN1Sequence.fromBytes(Uint8List bytes) : super.fromBytes(bytes) {
-    if (isPrimitive(tag) && !isSequence(tag)) {
+    if (isUniversal(tag) && !isSequence(tag)) {
       throw ASN1Exception('The tag $tag does not look like a sequence type');
     }
     _decodeSeq();