+++ test/app_test.dart
@@ -11,7 +11,7 @@ void main() {
   test('decode application tag', () {
     var bytes = Uint8List.fromList([113, 1, 0]);
 
-    var parser = ASN1Parser(bytes, encodeApplicationTagsAsObject: true);
+    var parser = ASN1Parser(bytes);
 
     while (parser.hasNext()) {
       var obj = parser.nextObject();